fix: 코드 리펙터링 및 디렉터리 수정
