+++ app/(auth)/_layout.tsx
@@ -1,15 +1,18 @@
 import * as NavigationBar from "expo-navigation-bar";
 import { Stack } from "expo-router";
 import { useEffect } from "react";
+import { Platform } from "react-native";
 
 export default function AuthLayout() {
 
   useEffect(() => {
+    if (Platform.OS !== "android") return;
     const hideNavigationBar = async () => {
       await NavigationBar.setVisibilityAsync("hidden");
-      await NavigationBar.setBehaviorAsync("overlay-swipe");
+      try {
+        await NavigationBar.setBehaviorAsync("overlay-swipe");
+      } catch {}
     };
-
     hideNavigationBar();
   }, []);
