+++ src/features/map/screens/MapScreen.tsx
@@ -1,0 +1,633 @@
+import CurrentLocationIcon from "@/assets/icons/current-location.svg";
+import { ClusterDetailItem, getClusterDetail } from "@/src/api/lights.api";
+import { getPassportDetail } from "@/src/api/passport/passport.api";
+import PassportDetail from "@/src/features/passport/screens/PassportDetail";
+import AddPlaceScreen from "@/src/features/place/screens/AddPlaceScreen";
+import {
+  NaverMapMarkerOverlay,
+  NaverMapView,
+} from "@mj-studio/react-native-naver-map";
+import * as Location from "expo-location";
+import { useCallback, useEffect, useRef, useState } from "react";
+import {
+  ActivityIndicator,
+  LogBox,
+  StyleSheet,
+  Text,
+  TouchableOpacity,
+  View,
+} from "react-native";
+import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { CenterPin } from "../components/CenterPin";
+import { ClusterBottomSheet } from "../components/ClusterBottomSheet";
+import { DiscoverBanner } from "../components/DiscoverBanner";
+import { LocationConfirmCard } from "../components/LocationConfirmCard";
+import { PassportPreviewCard } from "../components/PassportPreviewCard";
+import {
+  FrontClusterMarker,
+  PassportMarker,
+  ServerClusterMarker,
+  UserLocationMarker,
+} from "../components/markers/MapMarkers";
+import {
+  GREEN_MID,
+  INITIAL_BBOX,
+  TAB_BAR_HEIGHT,
+} from "../constants/mapConstants";
+import { useMapLights } from "../hooks/useMapLights";
+import {
+  FrontCluster,
+  PassportPreview,
+  PickedLocation,
+} from "../types/map.types";
+import {
+  clusterSingleMarkers,
+  dominantCategory,
+  naverReverseGeocode,
+} from "../utils/mapUtils";
+
+LogBox.ignoreLogs([
+  'Unsupported top level event type "topSvgLayout" dispatched',
+]);
+
+export default function MapScreen() {
+  const { bottom: safeBottom, top: safeTop } = useSafeAreaInsets();
+  const { lights, currentBBox, loadLights, handleCameraChanged } =
+    useMapLights();
+
+  const [showRegisterBtn, setShowRegisterBtn] = useState(true);
+  const [showAddPlace, setShowAddPlace] = useState(false);
+  const [userLocation, setUserLocation] = useState<{
+    latitude: number;
+    longitude: number;
+  } | null>(null);
+
+  const [previewVisible, setPreviewVisible] = useState(false);
+  const [previewData, setPreviewData] = useState<PassportPreview | null>(null);
+  const [previewLoading, setPreviewLoading] = useState(false);
+  const [previewError, setPreviewError] = useState<string | null>(null);
+
+  const [detailItem, setDetailItem] = useState<any>(null);
+  const [detailLoading, setDetailLoading] = useState(false);
+
+  const [pickingLocation, setPickingLocation] = useState(false);
+  const [showLocationConfirm, setShowLocationConfirm] = useState(false);
+  const [cameraCenter, setCameraCenter] = useState({
+    latitude: 37.5665,
+    longitude: 126.978,
+  });
+  const [pickedLocation, setPickedLocation] = useState<PickedLocation | null>(
+    null,
+  );
+  const [addressLoading, setAddressLoading] = useState(false);
+
+  const [activeClusterItems, setActiveClusterItems] = useState<
+    ClusterDetailItem[] | null
+  >(null);
+  const [discoverBanner, setDiscoverBanner] = useState<{
+    address: string | null;
+  } | null>(null);
+
+  const mapRef = useRef<any>(null);
+  const baseBottom = TAB_BAR_HEIGHT + safeBottom + 12;
+  const frontClusters = clusterSingleMarkers(lights, currentBBox);
+
+  useEffect(() => {
+    (async () => {
+      const { status } = await Location.requestForegroundPermissionsAsync();
+      if (status !== "granted") return;
+      const location = await Location.getCurrentPositionAsync({
+        accuracy: Location.Accuracy.Balanced,
+      });
+      const { latitude, longitude } = location.coords;
+      setUserLocation({ latitude, longitude });
+      setDiscoverBanner({ address: null });
+      const address = await naverReverseGeocode(latitude, longitude);
+      setDiscoverBanner({ address });
+    })();
+  }, []);
+
+  const onCameraChanged = useCallback(
+    (event: any) => {
+      const result = handleCameraChanged(event, pickingLocation);
+      if (result?.latitude && result?.longitude) {
+        setCameraCenter({
+          latitude: result.latitude,
+          longitude: result.longitude,
+        });
+      }
+    },
+    [handleCameraChanged, pickingLocation],
+  );
+
+  const handleSinglePinTap = async (item: any) => {
+    setPreviewVisible(true);
+    setPreviewData(null);
+    setPreviewError(null);
+    setPreviewLoading(true);
+    try {
+      const res = await getPassportDetail(item.passportId);
+      const d = res.data.data;
+      setPreviewData({
+        passportId: d.passportId,
+        spaceName: d.spaceName,
+        address: d.address,
+        visitedAt: d.visitedAt,
+        imageUrl: d.imageUrls?.[0] ?? null,
+        musicTitle: d.musicTitle ?? null,
+        musicArtist: d.musicArtist ?? null,
+        category: d.category ?? "ETC",
+      });
+    } catch {
+      setPreviewError("장소 정보를 불러오지 못했어요.");
+    } finally {
+      setPreviewLoading(false);
+    }
+  };
+
+  const handleServerClusterTap = async (item: any) => {
+    const lat = item.centerLatitude ?? item.latitude;
+    const lng = item.centerLongitude ?? item.longitude;
+    const expectedCount = item.count ?? 0;
+    const viewLatSpan = currentBBox
+      ? currentBBox.maxLat - currentBBox.minLat
+      : 0.1;
+    const viewLngSpan = currentBBox
+      ? currentBBox.maxLng - currentBBox.minLng
+      : 0.1;
+
+    try {
+      let content: ClusterDetailItem[] = [];
+      let divisor = 10;
+      for (let attempt = 0; attempt < 6; attempt++) {
+        const halfLat = viewLatSpan / divisor / 2;
+        const halfLng = viewLngSpan / divisor / 2;
+        const res = await getClusterDetail({
+          minLat: lat - halfLat,
+          maxLat: lat + halfLat,
+          minLng: lng - halfLng,
+          maxLng: lng + halfLng,
+          size: 50,
+        });
+        content = res.data.data?.content ?? [];
+        if (content.length <= expectedCount) break;
+        divisor *= 2;
+      }
+      if (content.length === 1) {
+        handleSinglePinTap({
+          ...content[0],
+          latitude: lat,
+          longitude: lng,
+          isCluster: false,
+        });
+      } else if (content.length > 1) {
+        setActiveClusterItems(content);
+      } else {
+        mapRef.current?.animateCameraTo({
+          latitude: lat,
+          longitude: lng,
+          zoom: 16,
+          duration: 400,
+        });
+      }
+    } catch {
+      mapRef.current?.animateCameraTo({
+        latitude: lat,
+        longitude: lng,
+        zoom: 16,
+        duration: 400,
+      });
+    }
+  };
+
+  const handleFrontClusterTap = async (cluster: FrontCluster) => {
+    if (cluster.items.length === 1) {
+      handleSinglePinTap(cluster.items[0]);
+      return;
+    }
+    const lats = cluster.items.map((it) => it.latitude);
+    const lngs = cluster.items.map((it) => it.longitude);
+    const pad = 0.0001;
+    try {
+      const res = await getClusterDetail({
+        minLat: Math.min(...lats) - pad,
+        maxLat: Math.max(...lats) + pad,
+        minLng: Math.min(...lngs) - pad,
+        maxLng: Math.max(...lngs) + pad,
+        size: 50,
+      });
+      const content: ClusterDetailItem[] = res.data.data?.content ?? [];
+      if (content.length > 0) setActiveClusterItems(content);
+    } catch {
+      setActiveClusterItems(
+        cluster.items.map((it) => ({
+          passportId: it.passportId,
+          thumbnailUrl: it.thumbnailUrl,
+          spaceName: it.spaceName,
+          districtCategory: it.districtCategory,
+          districtDisplayName: it.districtCategory,
+          category: it.category,
+          categoryDisplayName: it.category,
+          visitedAt: it.visitedAt,
+          visibility: "PUBLIC",
+          likeCount: it.likeCount,
+          scrapCount: it.scrapCount,
+        })),
+      );
+    }
+  };
+
+  const handleSheetItemSelect = (item: ClusterDetailItem) => {
+    setActiveClusterItems(null);
+    handleSinglePinTap({
+      ...item,
+      latitude: 0,
+      longitude: 0,
+      isCluster: false,
+    });
+  };
+
+  const handleOpenDetail = async () => {
+    if (!previewData) return;
+    setDetailLoading(true);
+    try {
+      const res = await getPassportDetail(previewData.passportId);
+      setDetailItem(res.data.data);
+      setPreviewVisible(false);
+      setPreviewData(null);
+    } catch (e) {
+      console.error("상세 조회 실패:", e);
+    } finally {
+      setDetailLoading(false);
+    }
+  };
+
+  const handleRegisterPress = () => {
+    setShowRegisterBtn(false);
+    setPickingLocation(true);
+    setPreviewVisible(false);
+  };
+
+  const handlePinLocation = async () => {
+    const { latitude, longitude } = cameraCenter;
+    setAddressLoading(true);
+    try {
+      const address = await naverReverseGeocode(latitude, longitude);
+      setPickedLocation({ latitude, longitude, address });
+    } finally {
+      setAddressLoading(false);
+      setShowLocationConfirm(true);
+    }
+  };
+
+  const handleLocationConfirm = () => {
+    setPickingLocation(false);
+    setShowLocationConfirm(false);
+    setShowAddPlace(true);
+  };
+
+  const handleLocationCancel = () => {
+    setShowLocationConfirm(false);
+    setPickedLocation(null);
+  };
+
+  const handleLocationPress = async () => {
+    const { status } = await Location.requestForegroundPermissionsAsync();
+    if (status !== "granted") {
+      alert("위치 권한이 필요합니다.");
+      return;
+    }
+    const location = await Location.getCurrentPositionAsync({
+      accuracy: Location.Accuracy.High,
+    });
+    const { latitude, longitude } = location.coords;
+    setUserLocation({ latitude, longitude });
+    mapRef.current?.animateCameraTo({
+      latitude,
+      longitude,
+      zoom: 16,
+      duration: 500,
+    });
+  };
+
+  return (
+    <View style={styles.container}>
+      <NaverMapView
+        ref={mapRef}
+        style={styles.map}
+        initialCamera={{ latitude: 37.5665, longitude: 126.978, zoom: 12 }}
+        isNightModeEnabled={true}
+        lightness={0}
+        isShowZoomControls={!pickingLocation}
+        onCameraChanged={onCameraChanged}
+      >
+        {userLocation && (
+          <NaverMapMarkerOverlay
+            latitude={userLocation.latitude}
+            longitude={userLocation.longitude}
+            anchor={{ x: 0.5, y: 0.5 }}
+            width={40}
+            height={40}
+          >
+            <UserLocationMarker />
+          </NaverMapMarkerOverlay>
+        )}
+
+        {lights
+          .filter((it) => it.isCluster)
+          .map((item, idx) => (
+            <NaverMapMarkerOverlay
+              key={`server-cluster-${idx}-${item.count ?? 0}`}
+              latitude={item.centerLatitude ?? item.latitude}
+              longitude={item.centerLongitude ?? item.longitude}
+              anchor={{ x: 0.5, y: 0.5 }}
+              width={56}
+              height={56}
+              onTap={() => handleServerClusterTap(item)}
+            >
+              <ServerClusterMarker
+                count={item.count ?? 0}
+                category={item.category}
+              />
+            </NaverMapMarkerOverlay>
+          ))}
+
+        {frontClusters.map((cluster) =>
+          cluster.items.length === 1 ? (
+            <NaverMapMarkerOverlay
+              key={`single-${cluster.items[0].passportId}`}
+              latitude={cluster.centerLat}
+              longitude={cluster.centerLng}
+              anchor={{ x: 0.5, y: 0.5 }}
+              width={56}
+              height={56}
+              onTap={() => handleFrontClusterTap(cluster)}
+            >
+              <PassportMarker
+                spaceName={cluster.items[0].spaceName}
+                category={cluster.items[0].category}
+              />
+            </NaverMapMarkerOverlay>
+          ) : (
+            <NaverMapMarkerOverlay
+              key={`fc-${cluster.id}-${cluster.items.length}`}
+              latitude={cluster.centerLat}
+              longitude={cluster.centerLng}
+              anchor={{ x: 0.5, y: 0.5 }}
+              width={64}
+              height={64}
+              onTap={() => handleFrontClusterTap(cluster)}
+            >
+              <FrontClusterMarker
+                count={cluster.items.length}
+                category={dominantCategory(cluster.items)}
+              />
+            </NaverMapMarkerOverlay>
+          ),
+        )}
+      </NaverMapView>
+
+      {pickingLocation && <CenterPin />}
+
+      {!pickingLocation && (
+        <View style={[styles.bottomBar, { bottom: baseBottom }]}>
+          <TouchableOpacity
+            style={styles.locationButton}
+            activeOpacity={0.8}
+            onPress={handleLocationPress}
+          >
+            <CurrentLocationIcon width={26} height={26} stroke="#000000" />
+          </TouchableOpacity>
+          {showRegisterBtn && (
+            <TouchableOpacity
+              style={styles.registerButton}
+              activeOpacity={0.85}
+              onPress={handleRegisterPress}
+            >
+              <Text style={styles.registerButtonText}>
+                나만의 장소 등록하기
+              </Text>
+            </TouchableOpacity>
+          )}
+        </View>
+      )}
+
+      {pickingLocation && !showLocationConfirm && (
+        <>
+          <View style={[styles.pickingGuideWrapper, { top: safeTop + 10 }]}>
+            <View style={styles.pickingGuideCard}>
+              <Text style={styles.pickingGuideText}>
+                지도를 움직여 핀을 원하는 위치에 맞추세요
+              </Text>
+            </View>
+          </View>
+          <View style={[styles.bottomBar, { bottom: baseBottom }]}>
+            <TouchableOpacity
+              style={styles.locationButton}
+              activeOpacity={0.8}
+              onPress={handleLocationPress}
+            >
+              <CurrentLocationIcon width={26} height={26} stroke="#000000" />
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={styles.registerButton}
+              activeOpacity={0.85}
+              onPress={handlePinLocation}
+              disabled={addressLoading}
+            >
+              {addressLoading ? (
+                <ActivityIndicator color={GREEN_MID} />
+              ) : (
+                <Text style={styles.registerButtonText}>이 위치로 고정</Text>
+              )}
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={styles.cancelButton}
+              activeOpacity={0.85}
+              onPress={() => {
+                setPickingLocation(false);
+                setShowRegisterBtn(true);
+              }}
+            >
+              <Text style={styles.cancelButtonText}>취소</Text>
+            </TouchableOpacity>
+          </View>
+        </>
+      )}
+
+      {discoverBanner && !pickingLocation && !previewVisible && (
+        <View style={[styles.discoveryWrapper, { bottom: baseBottom }]}>
+          <DiscoverBanner
+            address={discoverBanner.address}
+            onDismiss={() => setDiscoverBanner(null)}
+            onRegister={() => {
+              setDiscoverBanner(null);
+              setShowAddPlace(true);
+            }}
+          />
+        </View>
+      )}
+
+      {pickingLocation && showLocationConfirm && pickedLocation && (
+        <View style={[styles.discoveryWrapper, { bottom: baseBottom }]}>
+          <LocationConfirmCard
+            latitude={pickedLocation.latitude}
+            longitude={pickedLocation.longitude}
+            address={pickedLocation.address}
+            onConfirm={handleLocationConfirm}
+            onCancel={handleLocationCancel}
+          />
+        </View>
+      )}
+
+      {previewVisible && !pickingLocation && (
+        <View style={[styles.discoveryWrapper, { bottom: baseBottom + 60 }]}>
+          <PassportPreviewCard
+            preview={previewData}
+            loading={previewLoading}
+            error={previewError}
+            onClose={() => {
+              setPreviewVisible(false);
+              setPreviewData(null);
+            }}
+            onOpenDetail={handleOpenDetail}
+          />
+        </View>
+      )}
+
+      {detailLoading && (
+        <View style={styles.detailLoadingOverlay}>
+          <ActivityIndicator size="large" color={GREEN_MID} />
+        </View>
+      )}
+
+      {detailItem && (
+        <View style={StyleSheet.absoluteFill}>
+          <PassportDetail
+            item={detailItem}
+            editable={false}
+            onBack={() => {
+              setDetailItem(null);
+              setPreviewVisible(true);
+            }}
+          />
+        </View>
+      )}
+
+      {showAddPlace && (
+        <View style={StyleSheet.absoluteFill}>
+          <AddPlaceScreen
+            initialLatitude={pickedLocation?.latitude}
+            initialLongitude={pickedLocation?.longitude}
+            initialAddress={pickedLocation?.address ?? undefined}
+            onClose={() => {
+              setShowAddPlace(false);
+              setShowRegisterBtn(true);
+              setPickingLocation(false);
+              setPickedLocation(null);
+              loadLights(INITIAL_BBOX);
+            }}
+          />
+        </View>
+      )}
+
+      <ClusterBottomSheet
+        items={activeClusterItems}
+        onClose={() => setActiveClusterItems(null)}
+        onSelectItem={handleSheetItemSelect}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: { flex: 1 },
+  map: { flex: 1 },
+  detailLoadingOverlay: {
+    ...StyleSheet.absoluteFillObject,
+    backgroundColor: "rgba(255,255,255,0.7)",
+    justifyContent: "center",
+    alignItems: "center",
+    zIndex: 20,
+  },
+  pickingGuideWrapper: {
+    position: "absolute",
+    top: 0,
+    left: 16,
+    right: 16,
+    alignItems: "center",
+    zIndex: 10,
+  },
+  pickingGuideCard: {
+    backgroundColor: "rgba(16, 185, 129, 0.88)",
+    borderRadius: 12,
+    paddingVertical: 12,
+    paddingHorizontal: 20,
+  },
+  pickingGuideText: {
+    color: "white",
+    fontSize: 14,
+    fontWeight: "600",
+    textAlign: "center",
+  },
+  cancelButton: {
+    flex: 0.5,
+    backgroundColor: "white",
+    paddingVertical: 14,
+    borderRadius: 28,
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOpacity: 0.15,
+    shadowRadius: 8,
+    shadowOffset: { width: 0, height: 2 },
+    elevation: 5,
+  },
+  cancelButtonText: {
+    fontSize: 15,
+    fontWeight: "700",
+    color: "#EF4444",
+  },
+  bottomBar: {
+    position: "absolute",
+    left: 16,
+    right: 16,
+    flexDirection: "row",
+    alignItems: "center",
+    gap: 12,
+  },
+  locationButton: {
+    width: 48,
+    height: 48,
+    borderRadius: 24,
+    backgroundColor: "white",
+    justifyContent: "center",
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOpacity: 0.15,
+    shadowRadius: 6,
+    shadowOffset: { width: 0, height: 2 },
+    elevation: 4,
+  },
+  registerButton: {
+    flex: 1,
+    backgroundColor: "white",
+    paddingVertical: 14,
+    borderRadius: 28,
+    alignItems: "center",
+    shadowColor: "#000",
+    shadowOpacity: 0.15,
+    shadowRadius: 8,
+    shadowOffset: { width: 0, height: 2 },
+    elevation: 5,
+  },
+  registerButtonText: {
+    fontSize: 15,
+    fontWeight: "700",
+    color: "#1E293B",
+  },
+  discoveryWrapper: {
+    position: "absolute",
+    left: 16,
+    right: 16,
+  },
+});
